refactor: :art: wip

+++ server/src/invoice_reader/presenter.py
@@ -7,11 +7,13 @@ from invoice_reader import settings
 from invoice_reader.app.exceptions import (
     CLIENT_NOT_FOUND,
     EXISTING_CLIENT_EXCEPTION,
+    INVOICE_NOT_FOUND,
     MISSING_ENVIRONMENT_VARIABLE_EXCEPTION,
+    ROLLBACK,
     UNPROCESSABLE_FILE,
 )
 from invoice_reader.core import storage
-from invoice_reader.mappers.clients import ClientMapper
+from invoice_reader.mappers import ClientMapper, InvoiceMapper, UserMapper
 from invoice_reader.models import S3
 from invoice_reader.repository import (
     ClientRepository,
@@ -24,11 +26,11 @@ from invoice_reader.utils import logger, s3_utils
 LOGGER = logger.get_logger()
 
 
-def submit_invoice(
+def add_invoice(
     user_id: uuid.UUID,
     file: BinaryIO,
     filename: str | None,
-    invoice_data: invoice_schema.InvoiceCreate,
+    invoice_create: invoice_schema.InvoiceCreate,
     session: sqlmodel.Session,
 ) -> None:
     if not filename:
@@ -38,89 +40,97 @@ def submit_invoice(
     file_data = FileData(user_id=user_id, filename=filename)
     s3_model = S3.init(bucket=settings.S3_BUCKET_NAME)
     invoice_repository = InvoiceRepository(session=session)
+
+    invoice = InvoiceMapper.map_invoice_create_to_invoice(invoice_create)
     storage.store(
         file=file,
         file_data=file_data,
-        invoice_data=invoice_data,
+        invoice=invoice,
         invoice_repository=invoice_repository,
         s3_model=s3_model,
     )
 
 
-def register_user(user: user_schema.User, session: sqlmodel.Session) -> None:
-    user_repository = UserRepository(session=session)
-    user_repository.add(user)
-
-
 def get_user_by_email(email: str, session: sqlmodel.Session) -> user_schema.User | None:
     user_repository = UserRepository(session=session)
-    user = user_repository.get_user_by_email(email=email)
-    return user
+    user_model = user_repository.get_user_by_email(email=email)
+    return UserMapper.map_user_model_to_user(user_model) if user_model else None
 
 
-def add_user(user: user_schema.User, session: sqlmodel.Session) -> None:
+def add_user(user: user_schema.UserPresenter, session: sqlmodel.Session) -> None:
     user_repository = UserRepository(session=session)
-    user_repository.add(user=user)
+    user_model = UserMapper.map_user_to_model(user=user)
+    user_repository.add(user_model=user_model)
 
 
 def get_invoice(
-    user: user_schema.User, file_id: uuid.UUID, session: sqlmodel.Session
-) -> invoice_schema.InvoiceGetResponse:
+    user_id: uuid.UUID, file_id: uuid.UUID, session: sqlmodel.Session
+) -> invoice_schema.InvoiceResponse:
     invoice_repository = InvoiceRepository(session=session)
-    invoice_response = invoice_repository.get(user_id=user.user_id, file_id=file_id)
-    return invoice_response
+    invoice_model = invoice_repository.get(user_id=user_id, file_id=file_id)
+    if invoice_model:
+        invoice = InvoiceMapper.map_invoice_model_to_invoice(
+            invoice_model=invoice_model
+        )
+        return InvoiceMapper.map_invoice_to_response(invoice=invoice)
+    raise INVOICE_NOT_FOUND
 
 
 def get_paged_invoices(
-    user: user_schema.User, session: sqlmodel.Session, page: int, per_page: int
-) -> invoice_schema.PagedInvoiceGetResponse:
+    user_id: uuid.UUID, session: sqlmodel.Session, page: int, per_page: int
+) -> invoice_schema.PagedInvoiceResponse:
     invoice_repository = InvoiceRepository(session=session)
-    invoice_responses = invoice_repository.get_all(user_id=user.user_id)
+    invoice_models = invoice_repository.get_all(user_id=user_id)
     start = (page - 1) * per_page
     end = start + per_page
-    return invoice_schema.PagedInvoiceGetResponse(
+    invoices = InvoiceMapper.map_invoice_models_to_invoices(
+        invoice_models=invoice_models
+    )
+    return invoice_schema.PagedInvoiceResponse(
         page=page,
         per_page=per_page,
-        total=len(invoice_responses),
-        data=invoice_responses[start:end],
+        total=len(invoice_models),
+        data=InvoiceMapper.map_invoices_to_responses(invoices=invoices[start:end]),
     )
 
 
 def add_client(
-    user: user_schema.User, session: sqlmodel.Session, client: client_schema.Client
+    user_id: uuid.UUID,
+    session: sqlmodel.Session,
+    client_create: client_schema.ClientCreate,
 ) -> None:
     client_repository = ClientRepository(session=session)
     existing_client = client_repository.get_by_name(
-        user_id=user.user_id, client_name=client.client_name
+        user_id=user_id, client_name=client_create.client_name
     )
     if existing_client:
         raise EXISTING_CLIENT_EXCEPTION
-    client_repository.add(user_id=user.user_id, client=client)
-    LOGGER.info("New client added to the database.")
+    client = ClientMapper.map_client_create_to_client(client_create=client_create)
+    client_model = ClientMapper.map_client_to_model(client=client, user_id=user_id)
+    client_repository.add(client_model=client_model)
 
 
 def get_client(
-    user: user_schema.User, session: sqlmodel.Session, client_id: uuid.UUID
+    user_id: uuid.UUID, session: sqlmodel.Session, client_id: uuid.UUID
 ) -> client_schema.ClientResponse:
     client_repository = ClientRepository(session=session)
-    client_model = client_repository.get(user_id=user.user_id, client_id=client_id)
+    client_model = client_repository.get(user_id=user_id, client_id=client_id)
     if not client_model:
         raise CLIENT_NOT_FOUND
     client = ClientMapper.map_client_model_to_client(client_model=client_model)
-    client_response = ClientMapper.map_client_to_response(client)
+    client_response = ClientMapper.map_client_to_response(client=client)
     return client_response
 
 
 def get_paged_clients(
-    user: user_schema.User,
+    user_id: uuid.UUID,
     session: sqlmodel.Session,
     page: int,
     per_page: int,
 ) -> client_schema.PagedClientResponse:
     client_repository = ClientRepository(session=session)
-    clients = ClientMapper.map_client_models_to_clients(
-        client_repository.get_all(user_id=user.user_id, limit=per_page)
-    )
+    client_models = client_repository.get_all(user_id=user_id, limit=per_page)
+    clients = ClientMapper.map_client_models_to_clients(client_models=client_models)
     client_responses = ClientMapper.map_clients_to_responses(clients=clients)
     return client_schema.PagedClientResponse(
         page=page, per_page=per_page, total=len(clients), data=client_responses
@@ -133,10 +143,17 @@ def delete_invoice(
     if not settings.S3_BUCKET_NAME:
         raise MISSING_ENVIRONMENT_VARIABLE_EXCEPTION
     invoice_repository = InvoiceRepository(session=session)
-    invoice = invoice_repository.get(file_id=file_id, user_id=user_id)
+    invoice_model = invoice_repository.get(file_id=file_id, user_id=user_id)
+    if not invoice_model:
+        raise INVOICE_NOT_FOUND
+    invoice = InvoiceMapper.map_invoice_model_to_invoice(invoice_model=invoice_model)
     invoice_repository.delete(file_id=file_id, user_id=user_id)
+
+    # TODO: implement UOW and rollback
+    if not invoice.s3_path:
+        raise ROLLBACK
     s3 = S3.init(bucket=settings.S3_BUCKET_NAME)
-    suffix = s3_utils.get_suffix_from_s3_path(invoice.s3_path)
+    suffix = s3_utils.get_suffix_from_s3_path(s3_path=invoice.s3_path)
     s3.delete(suffix=suffix)
 
 
@@ -153,10 +170,14 @@ def delete_user(user_id: uuid.UUID, session: sqlmodel.Session) -> None:
 
 
 def update_invoice(
-    invoice_id: uuid.UUID, invoice: invoice_schema.Invoice, session: sqlmodel.Session
+    invoice_update: invoice_schema.InvoiceUpdate,
+    session: sqlmodel.Session,
 ) -> None:
     invoice_repository = InvoiceRepository(session=session)
-    invoice_repository.update(invoice_id=invoice_id, invoice=invoice)
+    invoice_model = InvoiceMapper.map_invoice_to_model(
+        InvoiceMapper.map_invoice_update_to_invoice(invoice_update=invoice_update)
+    )
+    invoice_repository.update(updated_invoice_model=invoice_model)
 
 
 def get_invoice_url(
@@ -166,9 +187,12 @@ def get_invoice_url(
         raise MISSING_ENVIRONMENT_VARIABLE_EXCEPTION
     s3 = S3.init(bucket=settings.S3_BUCKET_NAME)
     invoice_repository = InvoiceRepository(session=session)
-    invoice = invoice_repository.get(
-        file_id=invoice_id, user_id=user_id
-    )  # TODO: horrible! Refactor this s*** with proper mappers & schemas
+    invoice_model = invoice_repository.get(file_id=invoice_id, user_id=user_id)
+    if not invoice_model:
+        raise INVOICE_NOT_FOUND
+    invoice = InvoiceMapper.map_invoice_model_to_invoice(invoice_model=invoice_model)
+    if not invoice.s3_path:
+        raise ValueError(f"s3_path of invoice {invoice_id} not found.")
     suffix = s3_utils.get_suffix_from_s3_path(s3_path=invoice.s3_path)
     url = s3.create_presigned_url(suffix=suffix)
     return url
